[#78] refactor: RootView를 파일로 분리했어요.
