+++ T8-NoFrank/App/T8_NoFrankApp.swift
@@ -36,22 +36,3 @@ struct T8_NoFrankApp: App {
         }
     }
 }
-
-struct RootView: View {
-    @EnvironmentObject var router: AppRouter
-
-    var body: some View {
-        Group {
-            switch router.currentScreen {
-            case .home: HomeView()
-            case .turnOffAlarm: TurnOffAlarmView()
-            case .stonedust: StoneDustView()
-            }
-        }
-    }
-}
-
-#Preview {
-    RootView()
-        .environmentObject(AppRouter())
-}
